+++ web/src/components/vendors/VendorLinkCard.vue
